requirements-test.txt
pytest==8.3.4
pytest-cov==6.0.0
requests==2.32.5
paho-mqtt==2.1.0
Flask==3.1.2
Flask-CORS==6.0.2
psycopg2-binary==2.9.11
influxdb-client==1.50.0
minio==7.2.20
